% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drainage_basins_layer.R
\name{get_drainage_basins_layer}
\alias{get_drainage_basins_layer}
\title{Download World drainage basins spatial layer}
\usage{
get_drainage_basins_layer(path = ".", overwrite = FALSE, timeout = 60, ...)
}
\arguments{
\item{path}{a \code{character} of length 1. The path to the folder to save the
spatial layer. Note that a subdirectory \verb{basins/} will be created.}

\item{overwrite}{a \code{logical}. If \code{TRUE} it will override the existing files.
Default is \code{FALSE}.}

\item{timeout}{an \code{integer}. The timeout for downloading files.
Default is \code{60}. This number can be increased for slow Internet connection.}

\item{...}{other arguments to pass to \code{download.file()} (e.g. \code{headers}).}
}
\value{
No return value. Files will be written in \code{path}.
}
\description{
Downloads the spatial layer of drainage basins of the World available on the
Figshare website
(\url{https://figshare.com/collections/A_global_database_on_freshwater_fish_species_occurrences_in_drainage_basins/3739145}).
}
\examples{
\dontrun{
## Download drainage basins spatial layer ----
get_drainage_basins_layer(path = "data")
}
}
\references{
Tedesco P \emph{et al.} (2017) A global database on freshwater fish species
occurrence in drainage basins. Scientific Data, 4, 170141.
\url{https://doi.org/10.1038/sdata.2017.141}.
}
