% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marine_ecoregions_layer.R
\name{get_marine_ecoregions_layer}
\alias{get_marine_ecoregions_layer}
\title{Download marine ecoregions of the World (MEOW) spatial layer}
\usage{
get_marine_ecoregions_layer(path = ".", overwrite = FALSE, timeout = 60)
}
\arguments{
\item{path}{a \code{character} of length 1. The path to the folder to save the
spatial layer. Note that a subdirectory \verb{MEOW/} will be created.}

\item{overwrite}{a \code{logical}. If \code{TRUE} it will override the existing files.
Default is \code{FALSE}.}

\item{timeout}{an \code{integer}. The timeout for downloading files.
Default is \code{60}. This number can be increased for slow Internet connection.}
}
\value{
No return value. Files will be written in \code{path}.
}
\description{
Downloads the spatial layer of marine ecoregions of the World (MEOW)
available on the WWF website
(\url{https://www.worldwildlife.org/publications/marine-ecoregions-of-the-world-a-bioregionalization-of-coastal-and-shelf-areas}).
}
\examples{
\dontrun{
## Download marine ecoregions spatial layer ----
get_marine_ecoregions_layer(path = "data")
}
}
\references{
Spalding MD \emph{et al.} (2007) Marine Ecoregions of the World: A
Bioregionalization of Coastal and Shelf Areas. BioScience, 57(7), 573-583.
\url{https://doi.org/10.1641/B570707}.
}
