% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_marine.R
\name{ggmap_marine}
\alias{ggmap_marine}
\title{World map of a marine spatial metric}
\usage{
ggmap_marine(data, metric, title = NULL, palette = viridisLite::turbo(n = 100))
}
\arguments{
\item{data}{an \code{sf} object. The spatial layer (\code{POLYGONS}, \code{POINTS}, etc.)
containing the variable \code{metric} for which values are to be mapped. For
aesthetic purposes, this layer should contains only marine values.
See \code{ggmap_freshwater()} for mapping a freshwater (or terrestrial) layer.}

\item{metric}{a \code{character} of length 1. The column in \code{data} for which
values are to be mapped.}

\item{title}{a \code{character} of length 1. The title of the map (legend title).
Default is \code{NULL} (no title).}

\item{palette}{a \code{character} of colors used to categorize the values of
\code{metric}. Default is \code{viridisLite::turbo()} with 100 colors.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots a marine spatially-explicit metric on a World map. Basemap layers come
from Natural Earth website \url{https://www.naturalearthdata.com/}.
}
\examples{
## Add an example
}
