% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_freshwater.R
\name{ggmap_terrestrial}
\alias{ggmap_terrestrial}
\title{World map of a terrestrial spatial metric}
\usage{
ggmap_terrestrial(
  data,
  metric,
  title = NULL,
  palette = viridisLite::turbo(n = 100)
)
}
\arguments{
\item{data}{an \code{sf} object. The spatial layer (\code{POLYGONS}, \code{POINTS}, etc.)
containing the variable \code{metric} for which values are to be mapped. For
aesthetic purposes, this layer should contain only terrestrial (including
freshwater) values (marine areas will be masked by the ocean layer).
See \code{ggmap_marine()} for mapping a marine layer.}

\item{metric}{a \code{character} of length 1. The column in \code{data} for which
values are to be mapped. This variable must be numeric.}

\item{title}{a \code{character} of length 1. The title of the map (legend title).
Default is \code{NULL} (no title).}

\item{palette}{a \code{character} of colors used to categorize the values of
\code{metric}. Default is \code{viridisLite::turbo()} with 100 colors.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots a terrestrial spatially-explicit metric on a World map. Basemap layers
(continents, oceans, graticules) come from the Natural Earth website
\url{https://www.naturalearthdata.com/}. The map is plotted in the Robinson
projection (\code{EPSG = 54030}).
}
\examples{
## Attach packages ----
library("rphenofish")
library("ggplot2")

## Path to example marine layer ----
file_name <- system.file(file.path("extdata", "freshwater_fish_richness.gpkg"), 
                         package = "rphenofish")

## Import example freshwater layer ----
freshwater_fish_richness <- sf::st_read(file_name, quiet = TRUE)

freshwater_fish_richness

## Map freshwater fish richness ----
freshwater_map <- ggmap_terrestrial(data   = freshwater_fish_richness, 
                                    metric = "richness",
                                    title  = "Number of freshwater fish species")

freshwater_map

\dontrun{
## Save map in PNG ----
ggsave(plot     = freshwater_map, 
       filename = "freshwater_fish_richness.png",
       width    = 20, 
       height   = 13, 
       units    = "cm", 
       dpi      = 300, 
       bg       = "white")
}
}
