% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphenofish-package.R
\docType{package}
\name{rphenofish-package}
\alias{rphenofish}
\alias{rphenofish-package}
\title{rphenofish: An R Client to Access the PHENOFISH Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an interface to the PHENOFISH database biodiversity, morphology, behaviour and ecophysiology of World marine and freshwater fishes. This package allows to download and to handle PHENOFISH data. It is part of the FRB-CESAB working group PHENOFISH. \url{https://www.fondationbiodiversite.fr/en/the-frb-in-action/programs-and-projects/le-cesab/phenofish/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://phenofish.github.io/rphenofish}
  \item Report bugs at \url{https://github.com/phenofish/rphenofish/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicolas Loiseau \email{nicolas.loiseau@cnrs.fr} (\href{https://orcid.org/0000-0002-2469-1980}{ORCID})
}

}
\keyword{internal}
