% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_marine.R
\name{ggmap_marine}
\alias{ggmap_marine}
\title{Plot a spatial metric at the marine ecoregion level}
\usage{
ggmap_marine(data, metric, title = NULL, palette = viridisLite::turbo(n = 100))
}
\arguments{
\item{data}{an \code{sf} object with the values of \code{metric} for each MEOW polygon
(in rows).}

\item{metric}{a \code{character} of length 1. The column in \code{data} to map.}

\item{title}{a \code{character} of length 1. The title of the map (legend title).
Default is \code{NULL} (no title).}

\item{palette}{a \code{character} of colors used to categorize the values of
\code{metric}. Default is \code{viridisLite::turbo()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Represents a spatially-explicit metric on a World map at the marine
ecoregions scale.
}
\examples{
## Add an example
}
