% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_marine.R
\name{ggmap_marine}
\alias{ggmap_marine}
\title{World map of a marine spatial metric}
\usage{
ggmap_marine(data, metric, title = NULL, palette = viridisLite::turbo(n = 100))
}
\arguments{
\item{data}{an \code{sf} object. The spatial layer (\code{POLYGONS}, \code{POINTS}, etc.)
containing the variable \code{metric} for which values are to be mapped. For
aesthetic purposes, this layer should contains only marine values.
See \code{ggmap_freshwater()} for mapping a freshwater (or terrestrial) layer.}

\item{metric}{a \code{character} of length 1. The column in \code{data} for which
values are to be mapped.}

\item{title}{a \code{character} of length 1. The title of the map (legend title).
Default is \code{NULL} (no title).}

\item{palette}{a \code{character} of colors used to categorize the values of
\code{metric}. Default is \code{viridisLite::turbo()} with 100 colors.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots a marine spatially-explicit metric on a World map. Basemap layers come
from Natural Earth website \url{https://www.naturalearthdata.com/}. The map
is plotted in the Robinson projection (\code{EPSG = 54030}).
}
\examples{
## Path to example marine layer ----
file_name <- system.file(file.path("extdata", "marine_fish_richness.gpkg"), 
                         package = "rphenofish")

## Import example marine layer ----
marine_fish_richness <- sf::st_read(file_name, quiet = TRUE)

marine_fish_richness

## Map marine fish richness ----
marine_map <- ggmap_marine(data   = marine_fish_richness, 
                           metric = "richness",
                           title  = "Number of marine fish species")

marine_map

\dontrun{
## Attach ggplot2 package ----
library("ggplot2")

## Save map ----
ggsave(plot     = marine_map, 
       filename = "marine_fish_richness.png",
       width    = 20, 
       height   = 13, 
       units    = "cm", 
       dpi      = 300, 
       bg       = "white")
}
}
